changes to Counter int64

+++ internal/storage/storage.go
@@ -3,9 +3,9 @@ package storage
 type StorageInterface interface {
 	SetGauge(key string, value float64)
 	GetGauge(key string) (float64, bool)
-	SetCounter(key string, value float64)
-	GetCounter(key string) (float64, bool)
+	SetCounter(key string, value int64)
+	GetCounter(key string) (int64, bool)
 	Delete(key string)
 	GetAllGauges() map[string]float64
-	GetAllCounters() map[string]float64
+	GetAllCounters() map[string]int64
 }
